Fix clang warning (#62)

+++ src/device_consistency.c
@@ -650,7 +650,7 @@ int device_consistency_signature_list_sort_comparator(const void *a, const void 
         result = memcmp(signal_buffer_data(buf1), signal_buffer_data(buf2), len1);
     }
     else {
-        result = len1 - len2;
+        result = (int)(len1 - len2);
     }
 
     return result;
